requirements.txt
flask
stripe
resend
python-dotenv
gunicorn